Fixed #303, confirming password on password change and account deletion.

+++ src/routes/login/DeleteAccount.svelte
@@ -2,12 +2,15 @@
     import Spinning from '@components/app/Spinning.svelte';
     import Button from '@components/widgets/Button.svelte';
     import { Creator } from '@db/CreatorDatabase';
-    import type { User } from 'firebase/auth';
+    import { signInWithEmailAndPassword, type User } from 'firebase/auth';
     import TextField from '../../components/widgets/TextField.svelte';
     import validEmail from '../../db/isValidEmail';
     import { DB, locales } from '@db/Database';
     import Feedback from '@components/app/Feedback.svelte';
     import MarkupHtmlView from '@components/concepts/MarkupHTMLView.svelte';
+    import { auth } from '@db/firebase';
+    import isValidEmail from '../../db/isValidEmail';
+    import isValidPassword from './IsValidPassword';
 
     export let user: User;
 
@@ -15,20 +18,39 @@
 
     let deleteRequested = false;
     let confirmEmail: string;
+    let password = '';
     let deleteSubmitted = false;
     let successfullyDeleted: boolean | undefined = undefined;
+    let deleteFeedback: string | undefined = undefined;
 
     async function deleteAccount() {
+        if (auth === undefined) return;
+
         deleteSubmitted = true;
-        successfullyDeleted = await DB.deleteAccount();
-        return true;
+
+        const email = isValidEmail(confirmEmail)
+            ? confirmEmail
+            : Creator.usernameEmail(confirmEmail);
+        try {
+            await signInWithEmailAndPassword(auth, email, password);
+            successfullyDeleted = await DB.deleteAccount();
+        } catch (error) {
+            deleteFeedback = $locales.get(
+                (l) => l.ui.page.login.error.wrongPassword,
+            );
+        }
+        deleteSubmitted = false;
     }
 
-    function readyToDeleteAccount(email: string) {
+    function readyToDeleteAccount(email: string, pass: string) {
         const finalEmail = creator.isUsername()
             ? Creator.usernameEmail(email)
             : email;
-        return validEmail(finalEmail) && finalEmail === user.email;
+        return (
+            validEmail(finalEmail) &&
+            finalEmail === user.email &&
+            isValidPassword(pass)
+        );
     }
 </script>
 
@@ -50,7 +72,7 @@
 
         <form
             on:submit={() =>
-                readyToDeleteAccount(confirmEmail)
+                readyToDeleteAccount(confirmEmail, password)
                     ? deleteAccount()
                     : undefined}
         >
@@ -58,35 +80,47 @@
                 description={$locales.get((l) =>
                     creator.isUsername()
                         ? l.ui.page.login.field.username.description
-                        : l.ui.page.login.field.email.description
+                        : l.ui.page.login.field.email.description,
                 )}
                 placeholder={$locales.get((l) =>
                     creator.isUsername()
                         ? l.ui.page.login.field.username.placeholder
-                        : l.ui.page.login.field.email.placeholder
+                        : l.ui.page.login.field.email.placeholder,
                 )}
-                fill={true}
                 kind={creator.isUsername() ? undefined : 'email'}
                 bind:text={confirmEmail}
+                editable={!deleteSubmitted}
+            />
+            <TextField
+                kind="password"
+                description={$locales.get(
+                    (l) => l.ui.page.login.field.password.description,
+                )}
+                placeholder={$locales.get(
+                    (l) => l.ui.page.login.field.password.placeholder,
+                )}
+                bind:text={password}
+                editable={!deleteSubmitted}
+                validator={(pass) => isValidPassword(pass)}
             />
             <Button
                 background
                 submit
                 tip={$locales.get(
-                    (l) => l.ui.page.login.button.reallyDelete.tip
+                    (l) => l.ui.page.login.button.reallyDelete.tip,
                 )}
-                active={readyToDeleteAccount(confirmEmail)}
+                active={readyToDeleteAccount(confirmEmail, password)}
                 action={deleteAccount}
                 >{$locales.get(
-                    (l) => l.ui.page.login.button.reallyDelete.label
+                    (l) => l.ui.page.login.button.reallyDelete.label,
                 )}</Button
             >
-            {#if confirmEmail?.length >= 5 && !readyToDeleteAccount(confirmEmail)}
+            {#if confirmEmail?.length >= 5 && !readyToDeleteAccount(confirmEmail, password)}
                 <Feedback inline
                     ><MarkupHtmlView
                         inline
                         markup={$locales.get(
-                            (l) => l.ui.page.login.feedback.match
+                            (l) => l.ui.page.login.feedback.match,
                         )}
                     /></Feedback
                 >
@@ -105,11 +139,14 @@
     >
 {/if}
 
+{#if deleteFeedback}
+    <Feedback>{deleteFeedback}</Feedback>
+{/if}
+
 <style>
     form {
         display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
+        flex-direction: column;
         gap: var(--wordplay-spacing);
         margin: var(--wordplay-spacing);
     }
